requirements.txt
anyio==3.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
geographiclib==2.0
geojson==2.5.0
h11==0.14.0
h3==3.7.6
httpcore==0.16.3
httpx==0.23.3
idna==3.4
numpy==1.24.2
Pillow==9.5.0
psutil==5.9.5
pycparser==2.21
pyowm==3.3.0
PySocks==1.7.1
python-telegram-bot==20.2
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.4
speedtest-cli==2.1.3
urllib3==1.26.15
